Split CLI dependencies to multiple files
